page.js und dessen Ladezeiten waren offensichtlich nicht effizient. Umgestellt auf Data load in der page.svelte mit "loading" Kontrolle
